+++ ui/src/app/modules/settings/accessory-control-lists/accessory-control-lists.interfaces.ts
@@ -3,6 +3,8 @@ export interface Pairing {
   _isPaired?: boolean
   _username?: string
   _category?: string
+  _matter?: boolean
+  _matterOnly?: boolean
   name: string
   username: string
 }